requirements.txt
flask
flask-cors
pymongo
joblib
numpy
gunicorn
python-dotenv
